Sun, Nov 12, 2023, 6:34 AM +07:00

+++ cs50w/week7/project4/network/views.py
@@ -138,6 +138,8 @@ class api:
             return JsonResponse({"error": "Post doesn't exist"}, status=404)
 
         if request.method == "DELETE":
+            if post.user != request.user:
+                return JsonResponse({"error": "Unaothorize action"}, status=400)
             serialize = post.serialize()
             post.delete()
             return JsonResponse(serialize, safe=False)
